+++ CHANGELOG.md
@@ -5,3 +5,5 @@
 * Fixed TCP implementation: no false positives on losses
 * Added COBS encoding
 * Moved buffers inside frame handling structures
+* Removed length checking
+* Allow unlimited speed in StaticLimiter via `-b 0` (mostly useful for TCP)